Reorganizing the rules a bit.
