+++ src/module/rules.js
@@ -19,16 +19,16 @@ import get from './engine/common/transformer/get.js';
 import isCircumstanceBonus from './rules/conditions/is-circumstance.js';
 import isUntypedBonus from './rules/conditions/is-untyped.js';
 import logToConsole from './rules/actions/log.js';
-import calculateBaseAbilityModifier from './rules/actions/calculate-base-ability-modifier.js';
-import calculateBaseArmorClass from './rules/actions/calculate-base-armor-class.js';
-import calculateArmorModifiers from './rules/actions/calculate-armor-modifiers.js';
-import calculateBab from './rules/actions/calculate-bab.js';
-import calculateBaseSaves from './rules/actions/calculate-base-saves.js';
-import calculateSaveModifiers from './rules/actions/calculate-save-modifiers.js';
-import calculateCharacterLevel from './rules/actions/calculate-character-level.js';
-import calculateInitiative from './rules/actions/calculate-initiative.js';
-import calculateInitiativeModifiers from './rules/actions/calculate-initiative-modifiers.js';
-import calculateCmd from './rules/actions/calculate-cmd.js';
+import calculateBaseAbilityModifier from './rules/actions/actor/calculate-base-ability-modifier.js';
+import calculateBaseArmorClass from './rules/actions/actor/calculate-base-armor-class.js';
+import calculateArmorModifiers from './rules/actions/actor/calculate-armor-modifiers.js';
+import calculateBab from './rules/actions/actor/calculate-bab.js';
+import calculateBaseSaves from './rules/actions/actor/calculate-base-saves.js';
+import calculateSaveModifiers from './rules/actions/actor/calculate-save-modifiers.js';
+import calculateCharacterLevel from './rules/actions/actor/calculate-character-level.js';
+import calculateInitiative from './rules/actions/actor/calculate-initiative.js';
+import calculateInitiativeModifiers from './rules/actions/actor/calculate-initiative-modifiers.js';
+import calculateCmd from './rules/actions/actor/calculate-cmd.js';
 import calculateShipArmorClass from './rules/actions/starship/calculate-ac.js';
 import calculateShipCritThreshold from './rules/actions/starship/calculate-ct.js';
 import calculateDrift from './rules/actions/starship/calculate-drift.js';
