+++ src/module/rules/actions/actor/calculate-character-level.js
