+++ src/module/rules/actions/actor/calculate-cmd.js
